requirements.txt
psutil==5.7.0
hurry.filesize==0.9